Move CheckInput and SliceAlmostEqual
